Refactor tx helpers to a separate file
